+++ mindsdb/utilities/ml_task_queue/const.py
@@ -1,6 +1,11 @@
 from enum import Enum
 
 
+TASKS_STREAM_NAME = b'ml-tasks'
+TASKS_STREAM_CONSUMER_GROUP_NAME = 'ml_executors'
+TASKS_STREAM_CONSUMER_NAME = 'ml_executor'
+
+
 class ML_TASK_TYPE(Enum):
     LEARN = b'learn'
     PREDICT = b'predict'
